+++ gutenberg/packages/editor/src/hooks/index.js
@@ -1,3 +1,4 @@
+import './collaboration-conflict-block';
 import './custom-sources-backwards-compatibility';
 import './default-autocompleters';
 import './media-upload';